feat: split better the components
